fb: Extract fb_videomodes.h, shake fb.h a bit
